FEAT: issue api authorize

+++ jjapra/src/main/java/jjapra/app/controller/IssueController.java
@@ -15,6 +15,7 @@ import jjapra.app.service.IssueService;
 import jjapra.app.service.ProjectMemberService;
 import jjapra.app.service.ProjectService;
 import lombok.RequiredArgsConstructor;
+import org.apache.coyote.Response;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -81,14 +82,11 @@ public class IssueController {
     }
 
     @GetMapping("/issues/details/{id}")
-    public ResponseEntity<IssueDetailsResponse> getIssueDetails(@PathVariable("id") Integer id) {
+    public ResponseEntity<Issue> getIssueDetails(@PathVariable("id") Integer id) {
         Issue issue = issueService.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("Invalid issue Id:" + id));
-        List<Comment> comments = issueService.findCommentsByIssueId(id);
 
-        IssueDetailsResponse response = new IssueDetailsResponse(issue.getTitle(), issue.getDescription(), comments);
-        return ResponseEntity.status(HttpStatus.OK)
-                .body(response);
+        return ResponseEntity.status(HttpStatus.OK).body(issue);
     }
 
     @PostMapping("/comments/{issueId}")
@@ -105,19 +103,67 @@ public class IssueController {
     }
 
     @GetMapping("/projects/{projectId}/issues")
-    public List<Issue> getIssuesByProjectId(@PathVariable("projectId") Integer projectId) {
-        return issueService.findByProjectId(projectId);
+    public ResponseEntity<List<Issue>> getIssuesByProjectId(@PathVariable("projectId") Integer projectId,
+                                            @RequestHeader("Authorization") String token) {
+        Optional<Member> loggedInUser = jwtMember.getMember(token);
+        if (loggedInUser.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
+
+        if (loggedInUser.get().getRole().toString().equals("ADMIN")) {
+            return ResponseEntity.ok(issueService.findByProjectId(projectId));
+        }
+
+        if (projectMemberService.findByProjectAndMember(
+                projectService.findById(projectId).get(), loggedInUser.get()).isEmpty()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+        return ResponseEntity.ok(issueService.findByProjectId(projectId));
     }
 
     //issue 수정
     @PatchMapping("/issues/{id}")
-    public ResponseEntity<Issue> updateIssue(@PathVariable("id") Integer id, @RequestBody UpdateIssueRequest request) {
+    public ResponseEntity<Issue> updateIssue(@PathVariable("id") Integer id, @RequestBody UpdateIssueRequest request,
+                                             @RequestHeader("Authorization") String token) {
+        Optional<Member> loggedInUser = jwtMember.getMember(token);
+        if (loggedInUser.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
+
+        Optional<Issue> issue = issueService.findById(id);
+        if (issue.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
+
+        if (!loggedInUser.get().getRole().toString().equals("ADMIN") &&
+            projectMemberService.findByProjectAndMember(projectService.findById(issue.get().getProjectId()).get(),
+                    loggedInUser.get()).isEmpty()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+
         Issue updatedIssue = issueService.updateIssue(id, request);
         return ResponseEntity.ok(updatedIssue);
     }
+
     //issue 삭제
     @DeleteMapping("/issues/{id}")
-    public ResponseEntity<Void> deleteIssue(@PathVariable("id") Integer id) {
+    public ResponseEntity<Void> deleteIssue(@PathVariable("id") Integer id, @RequestHeader("Authorization") String token) {
+        Optional<Member> loggedInUser = jwtMember.getMember(token);
+        if (loggedInUser.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
+
+        Optional<Issue> issue = issueService.findById(id);
+        if (issue.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
+
+        if (!loggedInUser.get().getRole().toString().equals("ADMIN") &&
+            projectMemberService.findByProjectAndMember(projectService.findById(issue.get().getProjectId()).get(),
+                    loggedInUser.get()).isEmpty()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+
         issueService.deleteIssue(id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
